Cleaned up SparkViewNode, added ViewModel property to ViewDefinition.
