+++ src/FubuMVC.Spark/Registration/Nodes/SparkViewNode.cs
@@ -3,7 +3,6 @@ using FubuMVC.Core.Registration.Nodes;
 using FubuMVC.Core.Registration.ObjectGraph;
 using FubuMVC.Spark.Rendering;
 using FubuMVC.Spark.SparkModel;
-using Spark;
 
 namespace FubuMVC.Spark.Registration.Nodes
 {
@@ -19,7 +18,7 @@ namespace FubuMVC.Spark.Registration.Nodes
         protected override void configureObject(ObjectDef def)
         {
             var renderer = def.DependencyByType(typeof (IViewRenderer), typeof (ViewRenderer));
-            var viewDefinition = new ViewDefinition(createSparkDescriptor(true), createSparkDescriptor(false));
+            var viewDefinition = _descriptor.ToViewDefinition();
             
             var viewEntrySource =
                 renderer.DependencyByType(typeof (IRenderAction), typeof (RenderAction))
@@ -27,32 +26,6 @@ namespace FubuMVC.Spark.Registration.Nodes
                 .DependencyByType(typeof(IViewEntrySource), typeof(ViewEntrySource));
 
             viewEntrySource.DependencyByValue(viewDefinition);
-        }
-
-        private SparkViewDescriptor createSparkDescriptor(bool useMaster)
-        {
-            var sparkDescriptor = new SparkViewDescriptor().AddTemplate(_descriptor.ViewPath);
-            if (useMaster && _descriptor.Master != null)
-            {
-                appendMasterPage(sparkDescriptor, _descriptor.Master);
-            }
-
-            return sparkDescriptor;
-        }
-
-
-        private static void appendMasterPage(SparkViewDescriptor descriptor, ITemplate template)
-        {
-            if (template == null)
-            {
-                return;
-            }
-            descriptor.AddTemplate(template.ViewPath);
-            var viewDescriptor = template.Descriptor as ViewDescriptor;
-            if (viewDescriptor != null)
-            {
-                appendMasterPage(descriptor, viewDescriptor.Master);
-            }
         }
 
         public override string Description
